Disclosure Box Fix Attempt #1
So, I don't know which of the changes screwed up everything. My guess is that it was when I moved everything to the CSS file. However, for some reason it didn't register until like the 5th change. In addition, I had forgotten to connect the CSS file to the HTML file.